BUGFIX: Web: Organize Searches: Change Folder order
The comparison of retrieved project id for each folder was object
equivalence instead of comparing the int values by calling intValue().
Changed to
if ( projectId.intValue() != projectIdFromFolderId.intValue() ) {

+++ limelight_webapp/src/main/java/org/yeastrc/limelight/limelight_webapp/spring_mvc_parts/data_pages/rest_controllers/other_like_project/Project_OrganizeSearches_SetFoldersOrder_RestWebserviceController.java
@@ -153,7 +153,7 @@ public class Project_OrganizeSearches_SetFoldersOrder_RestWebserviceController {
 				if ( projectId == null ) {
 					projectId = projectIdFromFolderId;
 				} else {
-					if ( projectId != projectIdFromFolderId ) {
+					if ( projectId.intValue() != projectIdFromFolderId.intValue() ) {
 						//  Invalid request, searches not in from or to project
 						String msg = "Folders don't all have the same project id, projectFolderIdsInOrder: " + projectFolderIdsInOrder;
 						log.warn( msg );
